requirements.txt
streamlit==1.26.0
sentence-transformers==2.2.2
transformers==4.34.0
torch>=2.1.0
faiss-cpu==1.7.4
numpy==1.26.4
pickle5==0.0.12
